requirements.txt
fabric3>=1.13.1